NPC todos: - [x] Display background/subclass/etc in Statblock tab - [x] ~~(confirm if needed) Class styling in sidebar (copy from PC sheet Traits tab)~~ Not needed. Already covered. - [x] Show alignment in bio tab - [x] Show species in sidebar in Edit Mode, subtitle only in play mode - [x] Show size in sidebar in Edit Mode, subtitle only in play mode - [x] Make loyalty tracker match legendary trackers
Planned work for some additional missing items.

Generalized several components and context preparation functions so that NPCs and Characters can share.

Corrected Singleton warning loc key, as it has changed.

Added Creature Type to subtitle.
