+++ src/sheets/classic/Tidy5eActorSheetClassicV2Base.svelte.ts
@@ -1161,7 +1161,7 @@ export function Tidy5eActorSheetClassicV2Base<
         const singleton = dataModel?.metadata.singleton ?? false;
         if (singleton && this.actor.itemTypes[itemData.type].length) {
           ui.notifications.error(
-            game.i18n.format('DND5E.ActorWarningSingleton', {
+            game.i18n.format('DND5E.ACTOR.Warning.Singleton', {
               itemType: game.i18n.localize(
                 CONFIG.Item.typeLabels[itemData.type]
               ),
