+++ src/sheets/quadrone/actor/character-parts/traits/CharacterTraits.svelte
@@ -1,13 +1,12 @@
 <script lang="ts">
-  import SelectOptions from 'src/components/inputs/SelectOptions.svelte';
-  import SelectQuadrone from 'src/components/inputs/SelectQuadrone.svelte';
   import { FoundryAdapter } from 'src/foundry/foundry-adapter';
   import { getCharacterSheetQuadroneContext } from 'src/sheets/sheet-context.svelte';
-  import CharacterTraitClasses from './CharacterTraitClasses.svelte';
-  import CharacterTraitSpecies from './CharacterTraitSpecies.svelte';
-  import CharacterTraitBackground from './CharacterTraitBackground.svelte';
+  import ActorTraitClasses from '../../parts/ActorTraitClasses.svelte';
+  import ActorTraitSpecies from '../../parts/ActorTraitSpecies.svelte';
+  import ActorTraitBackground from '../../parts/ActorTraitBackground.svelte';
   import ActorTraitConfigurableListEntry from '../../parts/ActorTraitConfigurableListEntry.svelte';
   import { SpecialTraitsApplication } from 'src/applications-quadrone/special-traits/SpecialTraitsApplication.svelte';
+  import ActorTraitSize from '../../parts/ActorTraitSize.svelte';
 
   let context = $derived(getCharacterSheetQuadroneContext());
 
@@ -38,42 +37,14 @@
   </div>
 
   <div class="list traits">
-    <CharacterTraitClasses />
+    <ActorTraitClasses />
 
-    <CharacterTraitSpecies />
+    <ActorTraitSpecies />
 
-    <CharacterTraitBackground />
+    <ActorTraitBackground />
 
     <!-- Size -->
-    <div class="list-entry">
-      <div class="list-label">
-        <h4 class="font-weight-label">
-          {localize('DND5E.Size')}
-        </h4>
-      </div>
-      <div class="list-content">
-        <div class="list-values">
-          {#if context.unlocked}
-            <SelectQuadrone
-              document={context.actor}
-              field="system.traits.size"
-              value={context.system.traits.size}
-            >
-              <SelectOptions
-                data={context.config.actorSizes}
-                labelProp="label"
-              />
-            </SelectQuadrone>
-          {:else}
-            <ul class="pills">
-              <li class="pill pill-medium">
-                {context.size.label}
-              </li>
-            </ul>
-          {/if}
-        </div>
-      </div>
-    </div>
+    <ActorTraitSize />
 
     <!-- Speed -->
     <ActorTraitConfigurableListEntry
